requirements.txt
easyocr
opencv-python
Pillow
numpy
python-dateutil
